Add test: only copy one input file.
Some document types can contain several files. For instance HTML files
that link to local images and CSS files. We therefore copied all files
in a directory also if a single file was given as input. This behavior
is apparently dangerous, as complete directory trees had to be handled
this way, even if they were not relevant at all.

We should, instead, consider only the one input file. If a document
contains more than one file, it is still possible to ZIP these files
and pass in the ZIP file.